requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==3.1.7
black==24.8.0
click==8.1.7
ecdsa==0.19.0
fastapi==0.114.2
greenlet==3.1.0
h11==0.14.0
httptools==0.6.1
idna==3.10
mypy-extensions==1.0.0
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.9.1
pydantic_core==2.23.3
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.10
PyYAML==6.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typing==3.7.4.3
typing_extensions==4.12.2
uuid==1.30
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1